+++ src/module/fvtt-player-achievements.js
@@ -166,6 +166,9 @@ async function getAchivements(overrides) {
     if (achievement.image == "") {
       achievement.image = "modules/fvtt-player-achievements/images/default.webp";
     }
+    if (achievement.cloakedImage == "") {
+      achievement.cloakedImage = "modules/fvtt-player-achievements/images/default.webp";
+    }
   });
 
   const hideUnearnedAchievements = game.settings.get("fvtt-player-achievements", "hideUnearnedAchievements");
@@ -180,9 +183,11 @@ async function getAchivements(overrides) {
     } else if (cloakUnearnedAchievements) {
       retachievements = retachievements.map((achievement) => {
         if (!achievement.completedActors.includes(callingCharacterId)) {
-          achievement.title = "HIDDEN";
+          if (!achievement.showTitleCloaked) {
+            achievement.title = "HIDDEN";
+          }
           achievement.description = "HIDDEN";
-          achievement.image = "modules/fvtt-player-achievements/images/default.webp";
+          achievement.image = achievement.cloakedImage ?? "modules/fvtt-player-achievements/images/default.webp";
         }
         return achievement;
       });